Refactor the Device class
This puts the Remote class into the Device class to manage the remote
device in the same way that the QLowEnergyController manages the local
device.  The primary difference is that the Remote class only exists for
the duration of the connection, while the QLowEnergyController exists
for the lifetime of the Device class.

Signed-off-by: Ed Beroset <beroset@ieee.org>