Add everything required for heartrate sensor
This turns on and leaves on the heartrate sensor in the constructor of
the HeartRateService class, which is probably not very good for battery
life.  A better way to approach this might be to have a global timer
with a heartrate service and then read it via DBus from within here.

Signed-off-by: Ed Beroset <beroset@ieee.org>